Inital commit, demo app which uses swapi
express app with one endpoint `/endpoint` which fetches `id` of `type` from
swapi and optionally includes the films the entity is in.

Control of config is passed to Ambeeance which can then be used to modify the
values on the fly without app restart as well as persisting values across app
runs.